requirements.txt
torch>=2.1
numpy>=1.23
pandas>=1.5
pyyaml>=6.0
tensorboard>=2.12
datasets>=2.14
scikit-learn>=1.1
matplotlib>=3.6

